+++ packages/client/src/internals/TRPCAbortError.ts
@@ -1,0 +1,7 @@
+export class TRPCAbortError extends Error {
+  constructor() {
+    super('The operation was aborted.');
+    this.name = 'TRPCAbortError';
+    Object.setPrototypeOf(this, TRPCAbortError.prototype);
+  }
+}
