+++ packages/client/src/links/httpLink.ts
@@ -3,7 +3,7 @@ import { TRPCClientError } from '../TRPCClientError';
 import { transformRPCResponse } from '../internals/transformRPCResponse';
 import { httpRequest } from '../internals/httpRequest';
 import { HttpLinkOptions, TRPCLink } from './core';
-import { TRPCAbortError } from '../internals/TRPCAbortErrorSignal';
+import { TRPCAbortError } from '../internals/TRPCAbortError';
 
 export function httpLink<TRouter extends AnyRouter>(
   opts: HttpLinkOptions,
