+++ packages/client/src/links/wsLink.ts
@@ -168,9 +168,10 @@ export function createWSClient(opts: WebSocketClientOptions) {
         // gracefully replace old connection with this
         req.ws = activeConnection;
         closeIfNoPending(oldWs);
+        return;
       }
 
-      if (res.result.type === 'stopped') {
+      if (res.result.type === 'stopped' && conn === activeConnection) {
         req.callbacks.onDone?.();
       }
     };
